Move tests to test_dwpc.py
